[config] Init benchmark config
