+++ .gitignore
@@ -29,3 +29,4 @@ vendor
 a.out
 xk*.zip
 xkd.yml
+xkb.yml
